fix: separate schema functions from hooks
