+++ packages/apps/human-app/frontend/src/modules/signup/operator/utils/index.ts
@@ -1,2 +1,4 @@
 export * from './sort-form';
 export * from './staked-amount-formatter';
+export * from './add-stake-amount-call-arguments-schema';
+export * from './eth-kv-store-values-mutation-schema';
